feat: Droplist CT
